Checked the solution
I was supposed to use two classes to make the 3rd element the way it's
supposed to look. Originally I used one instead, but gave the two
classes same style settings in style.css. Now it is corrected to match
the exercise solution. (I was able to match the desired outcome with the
original setup, just not according to the instructions)